Parameters and log files for  ipyrad vs. Stacks
Contains the ipyrad parameters files used for the demultiplexation process of both assemblies used for comparison and the parameters for the remaining assembly steps in the case of ipyrad. It also contains the log files corresponding to Stacks' denovo_map.pl.
Note: The ipyrad's parameters files included here can also be found in the folder params files. They were included in both folders to avoid the necessity of shifting between folders according to the objective of file consultation.